requirements.txt
requests
pytest
requests-mock
google-genai
tzdata
